Introduce saturated categories
A saturated category is the standard mathematical definition of a category,
where every two matching morphisms can be composed to form another morphism.

The notion of an equivalence relation was factored out so that it could be
shared between categories and saturated categories.

Functions to convert between categories and saturated categories have been
defined, but the proofs of their existance leave a lot to be desired; I haven't
yet learned the skill of automated proving.

The concatenate function was rewritten so that it admits a more natural
induction (on the second argument rather than the first), because adding an edge
at the very end is more natural than adding one in the middle.  The original
concatenate was left in and renamed to concatenate', and proven equivalent to
concatenate.  (Again, the proof leaves much to be desired.)

A proof that the identity of a saturated category is unique up to morphism
equivalence was added.

Todo:
  * Add comments/documentation
  * Figure out automated proving and teach intuition/simpl that
      MorphismsEquivalent and PathsEquivalent are like '=' (or make a variant of
      inuition/simpl for them)
  * Figure out how to phrase and prove the theorem that, up to
      MorphismEquivalence and PathEquivalence, categories and saturated
      categories are the same thing, so I only have to prove theorems about one
      of them to get theorems about the other.
  * Define the opposite category
  * Define the category of functors from a given category
  * Define natural equivalence
  * Define limit and colimit
  * Define adjoint pairs and prove that (co)limit and the constant diagram
      functor are adjoint
  * Define full and faithful
  * Define retracts
  * Define (co)products, pushouts/pullbacks, sequential (co)limits as examples
      of particular types of (co)limits
  * Define initial and terminal objects
  * Define what it means to preserve limits and colimits
  * Define enriched categories?
      - Define 2-categories, n-categories?
  * Define the Yoneda embedding and prove the Yoneda lemma?
      - See if the Yoneda lemma can be extended to
          enriched categories/n-categories?