Introduce Saturated{Functor,NaturalTransformation}
Define functors and natrual transformations for saturated categories. Define
what it means for a morphism to be an isomorphism.  Define the identity functor.
Define natrual equivalences.

I attemped to define an identity natrual transformation, but got a bit stuck
in the proof.  Figuring out how to tell Coq to treat 'MorphismsEquivalent'
like '=' will fix this trivially.

The next steps are to define equivalence of categories, which is a functor
F with an ``inverse'' G such that FG and GF are natrually equivalent to the
identity functors.  Then I'll have enough machinary to state and prove one
direction of ``saturated categories and categories are equivalent things'';
it's that the category generated by unsaturating and then saturating is
equivalent to the original category.  Then I'll have to figure out the right
way to phrase the other direction, and then prove it.

I'll probably refactor categories and saturated categories into separate
definition files, and then make another file for proving that they're
equivalent.  It'd be nice to see if I can bootstrap the definitions for
functor and natural transformation on this equivalence, but I might have to do
it by hand.  Everything after that, though, will only have to be defined/proved
for either saturated categories or for categories, and not both, I hope.