Add associativity hints, partially define natural transformation composition
Some proofs are now marginally better automated with the help of some hints
about associativity in Category.v.

[ComposeFunctor] now has smarter implicit arguments.

Use [proj1_sig] rather than [match foo with _ => _ end].  This fixes the
problem with existential variables.  Using [proj2_sig] lets me finish fixing the
proof.  This construction (the inverse of a natural equivalence) is now somewhat
more automated, though the [Ltac]s are a bit slow.  Also, there's a kludge with
[?MorphismsEquivalent] where I really want to be using [MorphismsEquivalent].

I did some minor space clean-ups (I'm going to stick with no extra spaces at the
end of lines for future commits).

There are two ways of composing natural transformations, horizontally
([NTComposeF], because we're composing the functors), and vertically
([NTComposeT], because we're composing the natural transformations in the
obvious way).  [NTComposeT] went through fine, though it could do with slightly
more automation.  (Rather, I use [m] in the proof, but [m] is introduced by Coq.
This should be fixed.)  I'm having trouble with [NTComposeF], though, because
I can't get the type signatures to line up.

The next thing I want to do is say that natural equivalence is an equivalence
relation (with [Add Parametric Relation]), and that functor composition plays
nicely with this relation.  I then want to define two categories to be
naturally equivalent if there exist functors [F] and [G] where [FG] and [GF] are
naturally equivalent to the identity functor.

Perhaps I should pull the natural equivalence bits out in to a separate file
from the natural transformation bits.