Define a [GeneralizedEquivalenceRelation], almost define Cat
For [FunctorsEquivalent], we need a notion of [MorphismsEquivalent] that doesn't
require that both morphisms have the same domain and co-domain.  This is
[GeneralizedRelationsEquivalent] in EquivalenceRelation.v.  It's unfortunate
that [Add Parametric Relation] doesn't seem to support this generalized notion
of relations.  (And it complains about being "unable to manage a dependent
metavariable of higher-order type" if I try to define an equivalence relation
on [GeneralizedRelationsEquivalent] restricted to the case where the types match
up.)

The [InverseOf] lemmas in Category.v did not compile; a type annotation was
added to fix this.

The tactic [morphisms] is really slow.  I've un-automated a bit of
[CategoryIsomorphismComposition]; [firstorder] is slower than
[repeat (destruct 1)], and they both do the same thing in the beginning.  It's
also super-slow in [NaturalEquivalenceInverse], though I haven't touched that
yet.  We should probably figure out either how to make [morphisms 2] faster, or
find a more selective automation technique (that doesn't take 15-30 seconds) for
these two proofs.

A stricter notion of functor equivalence has been defined.  Two functors are
equivalent if, for all morphisms [m], they sent [m] to equivalent morphisms.
I'm having a tremendous amount of trouble with functor equivalence and
[GeneralizedRelationsEquivalent] in general.

I want a tactic that removes all instances of [GeneralizedRelationsEquivalent]
from both the hypotheses and goals, and replaces [a ~= b] with something like
[forall p, a = cast p b] where [cast A B (p : A = B) : A -> B] is a no-op.
Alternatively, if I can figure out how to reason about [~=] or [existT], then
[dependent destruct] and [inversion] might be sufficient.

For now, though, I've simply admitted a bunch of FunctorComposition lemmas.  The
first has a rather long commented out region of things I've tried. Help with
proving them would be appreciated.

The order of arguments on [ComposeFunctors] has been changed to match (I think)
the order of arguments on [Compose].  NaturalTransformation.v was modified
accordingly.

For any computable collection of Categories (i.e., any type [T] with a map
[f : T -> Category]) can now be made in to a Category of such objects.
ComputableCategory.v expresses this conversion.

If [Coercion] is sufficiently general, I intend to define a coercion from
[Functor] to a [Morphism] in any category of categories which includes its
domain and range.  This will give us nice things, like being able to use
[Compose] for [Functor]s.  It may also be the case that a functor between
categories of categories is a natural transformation, or is canonically
equivalent to one; I'll have to think on this more (or ask David).

Next up, math-wise, is defining what it means for two categories to be naturally
equivalent (and possibly proving that this is the same as them being isomorphic
in a category of categories, if that's true), and then defining the category
of functors from [C] to [D] for any categories [C] and [D].

After that, I'll probably either define [Colimit] and [Limit], or go back to
[Schema]s and define what it means for [Schema]s to be equivalent, so I can
show that Cat and Schema are equivalent categories.