Define a different variant of [FunctorsEquivalent], at Adam's suggestion
This one seems to have more hope than the last one, though I haven't been able
to prove as much with it, yet.  I wrote a bunch of [exists_assum_*] tactics
which (exhaustively) try [exists H] for all assumptions [H].  These dispense
with reflexivity and symmetry (where the final solving tactic is customized for
speed), but everything else is admitted.  I left everything that I was playing
with around in comments.