Define [Adjunction].
Use Awody's definition from "Category Theory" rather than the standard one,
because I had trouble defining Hom-functors:

As I understand it, Hom-functors go from [(OppositeCategory C) * C -> SetCat].
So they send pairs [(c', c)] to the set of morphisms between [c'] and [c].  So
far so good.

But the equivalence relation on morphisms in [CatSet], which are just functions
[f, g : A -> B] between sets, is [forall x : A, f x = g x].  But if we send
[(c', c)] to to the set of morphisms between [c'] and [c] (rather than
equivalence classes of morphisms), then equivalent morphisms are not equal, and
it's not the case that [Hom(-, -)] is a functor by our definitions.