Replace [MorphismsEquivalent] with [eq].
I made a mistake when defining [Category]; morphisms are equal, not equivalent.

I've now mostly fixed this mistake, cutting out a lot of code in the process.
Unfortunately, [t] and [intuition] now take longer to run, because there are
more things dealing with equality now.  Also [morphisms] seems to have mostly
broken.

I've defined [EquivalenceClassOf], mostly using Adam's code, in preparation
for fixing [saturate] (which is now broken).  However, I need to be able to say
that something is of type [EquivalenceClassOf foo], which I can't do now because
[EquivalenceClassOf foo] is of type [foo -> Type].

I've put a new tactic in Common.v that trys a given tactic after rewriting a
given hypothesis both ways ([rewrite] and [rewrite <-]).

[Functor]s and [NaturalTransformation]s are now also equal rather than
equivalent.

I've rewritten the [feq] tactic, but am still having some trouble getting things
like [f_equal] and [f_equal2] to work correctly with [Build_Functor] and even
more so with [Build_NaturalTransformation].

I've broken the definition of [FunctorCategory]; this is because I'm having
trouble with [f_equal] and [f_equal2].

I've broken, and just [admit]ted [PreComposeFunctorsNE] and
[PostComposeFunctorsNE].  [PreComposeFunctorsNE] has a proof that very nearly
works, and leaves me with two subgoals, both of which are the same and are
something like [?1 = x], and yet [eauto] fails to find a value for the
existential variable.  I don't know why.

There might be a better way to do proofs like [LeftIdentityFunctorNE]; it seems
like "prove that [f a = f b] because [a = b]" should have a name in core.

I've replaced [MorphismUnique] with the built-in [unique].