Define the limit and colimit functors.
As David suggested, the key to the proofs was uniqueness.

It would be nice to be able to use the extended proof for
[LimitFunctor_morphism_of], because they're eaiser to update if things change.
But that gives a bunch [let]s, and the definition of [LimitFunctor] doesn't go
through, and I didn't feel like spending the time to figure out how to make it
go through.  It would be nice if there were a way to tell Coq to
simplify/unfold/fold the [let]s into things like [projT1] and [projT2].