Define the Hom-set definition of Adjoint.
Use [unique] instead of [is_unique], because it's a "exists unique such that",
not "exists unique, and also it satisfies".

Prove that the three definitions are equivalent.  This was rather time-consuming
and I'm not sure that I fully understand why they're equivalent.  A decent chunk
of the proofs were done by type-matching and trial and error, though I think I
also learned some about adjunctions.

I've added a bunch of tactics to Common.v to simplify the proofs.  The largest
one is a tactic that finds all [proj1_sig]s in the goal and trys to rewrite
using the condition that they satisfy (the [proj2_sig] of the same thing).

Unfortunately, we can't use the hom-set adjoint definition to prove limit
functors, because that Hom.v and FunctorCategory.v together introduce universe
inconsistencies.