Define a special natural transformation between small categories.
It was very helpful to switch all morphisms to be [Set] and make the objects and
morphisms of [SmallCategory] [Set]s, so that I had a good check for when I went
wrong.  It feels conceptually cleaner, though, to allow all these things to be
types, especially for if Coq is eventually able to support universe
polymorphism.