Improve [SmallFunctor] and [SmallNaturalTransformation] reasoning.
We can now coerce between small and large things when the categories are of the
right type.  For example, since a [SmallFunctor] is just a [Functor] between
small categories (though it can live in [Set] rather than [Type], when things
are defined appropriately), if [C] and [D] are small categories, we can coerce
freely between [Functor C D] and [SmallFunctor C D], and I've proven the
relevant round-trip theorems in both directions.

I had to define [FunctorOnSmall] to get around the uniform inheritance
condition, so I've redefined convenience defninitions/lemmas that deal with
[FunctorOnSmall] automatically, and similarly for [NaturalTransformation].

I also cleaned up [SmallCat] a bit.