Define [Specialized] variants. Got Coq's fake universe poly. working!!!
[Specialized] variants take [Object] and [Morphism] types as parameters.  A bit
of careful work let me define a [Category] that lives at the same universe level
as its arguments (rather than at it's own, special universe level).  This means
that I actually have a good story for small categories and locally small
categories.  Most of the interface is almost the same; most of the proofs went
through unchanged.

As a test that this works, I built [FunctorCategory], and saw that the first of
the following checks gave [Set], and the second worked fine:
Parameter C D : @SmallSpecializedCategory nat (fun _ _ => unit).
Check (@SpecializedFunctor nat (fun _ _ => unit) nat (fun _ _ => unit)).
Check (SpecializedFunctorCategory C D : SmallSpecializedCategory _).

It's slightly annoying that I needed to define my own interface for each of the
record structures I built.  This is what it took to make my record objects live
at the same level as their arguments.  I'm still not sure why, though I have a
few guesses.  It's also slightly annoying that when defining new instances of
the objects, all the fields need to be primed.  I've defined [present_*] tactics
to (hopefully) get rid of the tics (') in the proofs that follow such
definitions.  See coq-club@ for a wierd problem that I had with [change] in a
tactic.

I've made [StructureEquality] to generalize the [*_eq_step_with] tactics that
are scattered throughout.

I'll probably define a [Category] wrapper for [SpecializedCategory] that hides
the objects and morphisms.  I should be able to define coercions back and forth
between [Category] and [SpecializedCategory], I hope.  I will probably use
[SpecializedCategory] in most places, except for where I need to talk about
a category of categories.