Many changes and improvements (hopefully)
I'm still not sure what the best divide is between [SpecializedCategory] and
[Category].  Is it better to use [SpecializedCategory] for only the low-level
stuff, or to try to do everything in [SpecializedCategory].

Try to make use of [Opaque]ness.  Not being able to tune it is a bit annoying.

Keep things folded in [CommaCategory] so that universal properties won't be so
slow.

Improve StructureEquality.v

Add a bunch of [Specialized] things.