Mostly implement left pushforward
It feels very much like the dual of right pushforward (limit -> colimit,
coslice -> slice, and then flip the orders of some arguments, and everything
goes through), so there should be a short way of doing it as the dual (or
something) of the right pushforward.  I'll have to ask David about it.