Generalize Grothendieck construction to Cat.
The functorial version of this (yet to come) will be the generalization that I
want (I think) of [sigT].  But I need the generalization of [forall] to express
this functorially, which I still don't have.  This construction should probably
get a better name.  http://ncatlab.org/nlab/show/Grothendieck+fibration says
that the category of elements of [F] is the oplax colimit of [F], so maybe I
should see what the (op?)lax limit of [F] is...