Work around https://github.com/HoTT/coq/issues/29
Matthieu says:
Hi Jason,
it is a tricky issue due to the fact that the system does not check constraints
but still infers them, I'm working on changing that (which will open the ability
to have named universes in the future as well). In the meantime, you can remove
the {Category objC} arguments to src and dst and it will go through.